Refactored Code to Multiple Files including header
