+++ LRU Cache Implementation/main.c
@@ -1,0 +1,13 @@
+#include "lru.h"
+
+int main() {
+    HashNode** hashMap = NULL;
+    Queue* queue = NULL;
+
+    displayMenuInstructions();
+
+    while(true)
+    {
+        handleCommand(&hashMap, &queue);
+    }
+}
